refactor: move queries out of routes
